refactor: 불필요한 코드 삭제 및 코드 중복 해결
